[CP-2163] Refactor: Move common logic to separate file as useNoNewDevicesDetectedHook
